requirements.txt
numpy
opencv-python
pillow
customtkinter